TASK 3 (interactive_colorization_model)/requirements-file.txt
torch==2.4.0
torchvision==0.19.0
numpy==1.24.3
matplotlib==3.7.1
Pillow==9.4.0
opencv-python==4.8.0
PyQt5==5.15.9
